# docs/openapi/velontri_openapi.yaml
openapi: 3.1.0
info:
  title: Velontri API
  version: 1.0.0
  description: |
    Complete API for all 14 Velontri microservices.
    All responses use a standard envelope: { success, message, data, meta }
    All errors use: { success: false, error: { code, message, field }, request_id }
  contact:
    name: Velontri Engineering
    email: dev@velontri.com
servers:
  - url: https://api.velontri.com/api/v1
    description: Production
  - url: https://api.staging.velontri.com/api/v1
    description: Staging
  - url: http://localhost:8001/api/v1
    description: Auth service (local)
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: RS256 JWT. Access token TTL is 15 minutes.
...
paths:
...
  /wallet/balance:
    get:
      tags: [Wallet]
      summary: Get wallet balance and rewards
      security: [{ BearerAuth: [] }]
      responses:
        '200':
          description: Wallet balance
          content:
            application/json:
              example:
                data:
                  balance: "125000.00"
                  held_balance: "50000.00"
                  available_balance: "75000.00"
                  rewards_points: 2500
                  currency: "NGN"
  /wallet/topup:
    post:
      tags: [Wallet]
      summary: Credit wallet from external payment
      security: [{ BearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [amount, payment_reference]
              properties:
                amount:            { type: number, example: 50000 }
                currency:          { type: string, example: "NGN" }
                payment_reference: { type: string, format: uuid, example: "pay-uuid-here" }
      responses:
        '201': { description: Wallet credited }
  /wallet/withdraw:
    post:
      tags: [Wallet]
      summary: Initiate a withdrawal
      security: [{ BearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [amount, bank_code, bank_account_number]
              properties:
                amount:              { type: number, example: 25000 }
                bank_code:           { type: string, example: "058" }
                bank_account_number: { type: string, example: "0123456789" }
                description:         { type: string }
      responses:
        '200': { description: Withdrawal initiated }
  /wallet/transfer:
    post:
      tags: [Wallet]
      summary: Transfer to another user's wallet
      security: [{ BearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [recipient_user_id, amount]
              properties:
                recipient_user_id: { type: string, format: uuid }
                amount:            { type: number, example: 10000 }
                description:       { type: string }
      responses:
        '200': { description: Transfer completed }
  /wallet/transactions:
    get:
      tags: [Wallet]
      summary: Paginated transaction history
      security: [{ BearerAuth: [] }]
      parameters:
        - { name: page,      in: query, schema: { type: integer, default: 1 } }
        - { name: page_size, in: query, schema: { type: integer, default: 20 } }
      responses:
        '200': { description: Transaction list }
  /wallet/rewards/redeem:
    post:
      tags: [Wallet]
      summary: Redeem rewards points for wallet credit
      security: [{ BearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [points]
              properties:
                points: { type: integer, minimum: 1, example: 500 }
      responses:
        '200': { description: Points redeemed }

  # ── PAYMENTS ─────────────────────────────────────────────────────────────────
...